// frontend/src/app/core/models/site-info.model.ts
export interface SiteInfo {
  _id?: string;
  brandName?: string;
  heroTitle?: string;
  heroDesc?: string;
  aboutTitle?: string;
  aboutDesc1?: string;
  aboutDesc2?: string;
  academicTitle?: string;
  academicDesc?: string;
  philosophyTitle?: string;
  philosophyDesc?: string;
  githubUrl?: string;
  linkedinUrl?: string;
  profileImage?: string;
  logoImage?: string;
  brandDesc?: string;
  contactTitle?: string;
  contactSubtitle?: string;
  contactEmail?: string;
  emailProvider?: string;
  emailjsServiceId?: string;
  emailjsTemplateId?: string;
  emailjsPublicKey?: string;
  formspreeUrl?: string;
  web3formsAccessKey?: string;
  createdAt?: string;
  updatedAt?: string;
}